[FIX][OBJ]Enabled 64bit configurations to properly compile with the correct types (tested). * tlocBasicTypes         - Since _WIN32 is defined in both win32 and x64 builds, the correct           types were not being defined, causing tl_int to be defined incorrectly           as a __int32 as well as others in the x64 build. Had to swap the #if           statement to allow for the proper typedefs to be called. * tlocTesting         - The Release_StdAlloc build was missing USING_STD_ALLOC preprocessor           define, which caused a runtime error in the x64 build.         - Since x64 now has proper types for tl_int, the remaining s32           references in various tests had to be changed to tl_int, in order to           prevent the the compiler from warning of data loss between an s32 and           tl_int.         - Various variables were of type tl_int, and were being compared with           tl_size variables. Had to change these to tl_size as well to prevent           signed unsigned mismatch warnings. * tlocMemoryPool         - static casted all size_type inputs, to the int_type (tl_int) when           assigning the size_type input into one of the int_type member           variables. * tlocAlgorithms         - in shuffle, a difference_type variable was being used when an int_type           (provided by the generator) was being expected when getting a random           integer. Static casted it to an int_type. In the other shuffle           function, an int_type does not have to be provided by the user so the           difference_type is casted to a tl_size, and a comment was modified,           asking that the function provided by the user only has a single input           with type tl_size.

+++ proj/VS/2008/tlocTesting/tlocInputTest.cpp
@@ -258,7 +258,8 @@ namespace TestingInput
       sampleInputMouse<Mouse<> > callback(mouse);
       mouse->Register(&callback);
 
-      SendMousePress(a_axis, a_data, a_x, a_y);
+      SendMousePress(a_axis, a_data, 
+                     static_cast<LONG>(a_x), static_cast<LONG>(a_y));
 
       while (countDown.ElapsedMilliSeconds() < 1000 &&
              callback.m_movementEvents == 0)
